+++ Tombolo/server/jobs/lzMonitoring/lzFileMovementMonitoring.js
@@ -1,0 +1,379 @@
+const models = require('../../models');
+const { parentPort } = require('worker_threads');
+const { decryptString } = require('../../utils/cipher');
+const { FileSprayService } = require('@hpcc-js/comms');
+const { getClusterOptions } = require('../../utils/getClusterOptions');
+const path = require('path');
+const {} = require('./lzMonitoringUtils');
+const { generateNotificationId } = require('../jobMonitoring/monitorJobsUtil');
+
+const {
+  landingZoneMonitoring: LandingZoneMonitoring,
+  cluster: Cluster,
+  monitoring_types: MonitoringTypes,
+  notification_queue: NotificationQueue,
+  asr_products: AsrProducts,
+} = models;
+
+// Recursively list all files (no name filter)
+async function listFilesRecursively({
+  lzFileMovementMonitoringId,
+  fss,
+  dropzone,
+  netaddr,
+  currentPath,
+  depth,
+  maxDepth,
+  pattern,
+}) {
+  if (depth > maxDepth) return [];
+  const resp = await fss.FileList({
+    DropZoneName: dropzone,
+    Netaddr: netaddr,
+    Path: currentPath,
+    directoryOnly: false,
+  });
+  const entries = resp?.files?.PhysicalFileStruct || [];
+  let results = [];
+  for (const e of entries) {
+    if (e.isDir) {
+      // descend into sub-directory
+      const subPath = path.posix.join(currentPath, e.name, '/');
+      const subFiles = await listFilesRecursively({
+        lzFileMovementMonitoringId,
+        fss,
+        dropzone,
+        netaddr,
+        currentPath: subPath,
+        depth: depth + 1,
+        maxDepth,
+        pattern,
+      });
+      results = results.concat(subFiles);
+    } else {
+      // record file only if it matches the wildcard pattern
+      if (wildcardMatch(pattern, e.name)) {
+        results.push({
+          lzFileMovementMonitoringId,
+          fileName: e.name,
+          directory: currentPath,
+          modifiedtime: e.modifiedtime,
+        });
+      }
+    }
+  }
+  return results;
+}
+
+// Match strings using * and ? wildcards
+function wildcardMatch(pattern, str) {
+  const escaped = pattern.replace(/[-\\/^$+?.()|[\]{}]/g, '\\$&');
+  const regex = new RegExp(
+    '^' + escaped.replace(/\*/g, '.*').replace(/\?/g, '.') + '$'
+  );
+  return regex.test(str);
+}
+
+// Local time at cluster
+function findLocalDateTimeAtCluster(timeZoneOffset) {
+  const newDate = new Date();
+  const localTime = new Date(newDate.getTime() + timeZoneOffset * 60 * 1000);
+  return localTime;
+}
+
+const monitoring_name = 'Landing Zone Monitoring';
+
+(async () => {
+  // Get monitoring type ID for "Job Monitoring"
+  const { id } = await MonitoringTypes.findOne({
+    where: { name: monitoring_name },
+    raw: true,
+  });
+  if (!id) {
+    throw new Error(`Monitoring type ${monitoring_name} not found.`);
+  }
+  const monitoringTypeId = id;
+
+  // Start time
+  const startTime = new Date().getTime();
+  try {
+    parentPort.postMessage({
+      level: 'info',
+      text: 'Landing Zone (file movement) Monitoring started',
+    });
+
+    // Get all lz monitoring with isActive = true and approvalStatus = approved
+    const activeLzMonitorings = await LandingZoneMonitoring.findAll({
+      where: {
+        isActive: true,
+        approvalStatus: 'approved',
+        lzMonitoringType: 'fileMovement',
+      },
+
+      include: [
+        {
+          model: Cluster,
+          required: true,
+          attributes: [
+            'id',
+            'thor_host',
+            'thor_port',
+            'username',
+            'hash',
+            'allowSelfSigned',
+            'timezone_offset',
+          ],
+        },
+      ],
+      raw: true,
+      nest: true,
+    });
+
+    // If 0 activeLzMonitorings, log and return
+    if (activeLzMonitorings.length === 0) {
+      parentPort.postMessage({
+        level: 'verbose',
+        text: 'No active landing zone (file movement) monitoring found',
+      });
+      return;
+    }
+
+    // Iterate over activeLzMonitorings, if none of them are lzMonitoringType = 'fileMovement' log and return
+    const uniqueClusters = [];
+    activeLzMonitorings.forEach(lzMonitoring => {
+      // Iterate over uniqueCluster arry and make sure lzMonitoring.clusterId is unique in there
+      let isClusterIdUnique = true;
+      uniqueClusters.forEach(c => {
+        if (c.id == lzMonitoring.clusterId) {
+          isClusterIdUnique = false;
+        }
+      });
+
+      if (isClusterIdUnique) {
+        let clusterPw;
+        if (lzMonitoring.cluster.hash) {
+          clusterPw = decryptString(lzMonitoring.cluster.hash);
+        }
+        uniqueClusters.push({ ...lzMonitoring.cluster, password: clusterPw });
+      }
+    });
+
+    parentPort.postMessage({
+      level: 'verobse',
+      text: `${activeLzMonitorings.length} active landing zone monitoring(s) tracking file movement`,
+    });
+
+    // create uniqueClustersObj for easy access
+    const uniqueClustersObj = {};
+    uniqueClusters.forEach(c => {
+      uniqueClustersObj[c.id] = c;
+    });
+
+    // Iterate over activeLzMonitorings and make async place holder calls
+    const allMatchedFiles = [];
+    const filesThatPassedThreshold = [];
+
+    for (const monitoring of activeLzMonitorings) {
+      try {
+        const {
+          id: lzFileMovementMonitoringId,
+          clusterId,
+          metaData: { monitoringData },
+        } = monitoring;
+
+        // Destructure monitoringData further
+        const { dropzone, machine, directory, maxDepth, fileName } =
+          monitoringData;
+
+        // Get all the files in the directory plus all subdirectories until maxDepth
+        const fss = new FileSprayService(
+          getClusterOptions(
+            {
+              baseUrl: `${uniqueClustersObj[clusterId].thor_host}:${uniqueClustersObj[clusterId].thor_port}`,
+              userID: uniqueClustersObj[clusterId].username || '',
+              password: uniqueClustersObj[clusterId].password || '',
+            },
+            uniqueClustersObj[clusterId].allowSelfSigned
+          )
+        );
+
+        // get only matching files
+        const matches = await listFilesRecursively({
+          lzFileMovementMonitoringId,
+          fss,
+          dropzone,
+          netaddr: machine,
+          currentPath: directory,
+          depth: 1,
+          maxDepth,
+          pattern: fileName,
+        });
+
+        allMatchedFiles.push(...matches);
+      } catch (error) {
+        parentPort.postMessage({
+          level: 'error',
+          text: `Error while getting files from File Spray service: ${error.message}`,
+        });
+      }
+    }
+
+    // If no files found, log and return
+    if (allMatchedFiles.length === 0) {
+      parentPort.postMessage({
+        level: 'verbose',
+        text: 'Landing zone (file movement) monitoring did not find any matching file in specified landing zones',
+      });
+      return;
+    }
+
+    // For ease of access, create lzFileMovementMonitoringAsObj with id
+    const lzFileMovementMonitoringAsObj = {};
+    activeLzMonitorings.forEach(lzMonitoring => {
+      lzFileMovementMonitoringAsObj[lzMonitoring.id] = lzMonitoring;
+    });
+
+    // Iterate over allMatchedFiles and evaluate whether the file has passed the threshold
+    for (let f of allMatchedFiles) {
+      const { lzFileMovementMonitoringId, modifiedtime } = f;
+      const cid =
+        lzFileMovementMonitoringAsObj[lzFileMovementMonitoringId].clusterId;
+      const threshold =
+        lzFileMovementMonitoringAsObj[lzFileMovementMonitoringId].metaData
+          .monitoringData.threshold;
+      const clusterOffset = uniqueClustersObj[cid].timezone_offset;
+      const localTimeAtCluster = findLocalDateTimeAtCluster(clusterOffset);
+      // compute age of file in ms and compare to threshold (minutes)
+      const fileDate = new Date(modifiedtime);
+      const ageMs = localTimeAtCluster - fileDate;
+      const thresholdMs = threshold * 60 * 1000;
+      if (ageMs > thresholdMs) {
+        filesThatPassedThreshold.push({
+          ...f,
+          ageInMins: Math.round(ageMs / 60000, 0),
+        });
+      }
+    }
+
+    // If FilesThatPassedThreshold is empty, log and return
+    if (filesThatPassedThreshold.length === 0) {
+      parentPort.postMessage({
+        level: 'verbose',
+        text: 'Landing zone (file movement) monitoring did not find any file that passed the threshold',
+      });
+      return;
+    }
+    // Queue notification for each file that passed the threshold
+    parentPort.postMessage({
+      level: 'info',
+      text: `Queing notification for ${filesThatPassedThreshold.length} file(s) that passed the threshold`,
+    });
+
+    for (const file of filesThatPassedThreshold) {
+      try {
+        const {
+          fileName,
+          lzFileMovementMonitoringId,
+          ageInMins,
+          directory,
+          modifiedtime,
+        } = file;
+
+        const {
+          clusterId,
+          applicationId,
+          metaData: {
+            asrSpecificMetaData,
+            contacts: {
+              primaryContacts,
+              secondaryContacts = [],
+              notifyContacts = [],
+            },
+            monitoringData: { dropzone, machine, threshold },
+          },
+        } = lzFileMovementMonitoringAsObj[lzFileMovementMonitoringId];
+
+        // If asr related the notification ID must be prefixed with the asr product short code
+        let notificationPrefix = 'LZFM';
+        if (asrSpecificMetaData?.productCategory) {
+          try {
+            const { productCategory } = asrSpecificMetaData;
+            // Geet product category
+            const asrProduct = await AsrProducts.findOne({
+              where: { id: productCategory },
+              attributes: ['shortCode'],
+              raw: true,
+            });
+            notificationPrefix = asrProduct.shortCode;
+          } catch (error) {
+            parentPort.postMessage({
+              level: 'warn',
+              text: `Error while getting ASR product category: ${error.message}`,
+            });
+          }
+        }
+
+        let notificationId = generateNotificationId({
+          notificationPrefix,
+          timezoneOffset: uniqueClustersObj[clusterId].timezone_offset,
+        });
+
+        // issue object
+        const issue = {
+          title: `File ${fileName} is stuck at ${directory} for ${ageInMins} minutes`,
+          cluster: uniqueClustersObj[clusterId].name,
+          dropzone,
+          machine,
+          directory,
+          fileModifiedTime: modifiedtime,
+          threshold,
+          fileAge: ageInMins,
+        };
+
+        // Add to notification queue
+        await NotificationQueue.create({
+          type: 'email',
+          notificationOrigin: 'Landing Zone Monitoring',
+          originationId: monitoringTypeId,
+          deliveryType: 'immediate',
+          templateName: 'lzFileMovementMonitoring',
+          notificationOrigin: 'Landing Zone Monitoring',
+          metaData: {
+            applicationId,
+            notificationId,
+            subject: `${fileName} is stuck at ${directory} for ${ageInMins} minutes`,
+            mainRecipients: primaryContacts,
+            cc: [...secondaryContacts, ...notifyContacts],
+            notificationDescription: `File ${fileName} is stuck at ${directory} for ${ageInMins} minutes. The maximum threshold is ${threshold} minutes.`,
+            asrSpecificMetaData,
+            issue,
+            remedy: {
+              instruction: `Check the file ${fileName} at ${directory} on ${machine} and ensure it is processed or removed from the landing zone.`,
+            },
+            firstLogged: new Date(),
+          },
+          createdBy: 'System',
+        });
+      } catch (error) {
+        parentPort.postMessage({
+          level: 'error',
+          text: `Error while queuing notification for file ${file.fileName}: ${error.message}`,
+        });
+      }
+    }
+  } catch (error) {
+    // Log error
+    parentPort.postMessage({
+      level: 'error',
+      text: `Error while monitoring landing zone for file movement: ${error.message}`,
+    });
+  } finally {
+    // End time
+    const endTime = new Date().getTime();
+    const duration = endTime - startTime;
+    parentPort.postMessage({
+      level: 'info',
+      text: `Landing Zone (file movement) Monitoring completed in ${duration} ms`,
+    });
+  }
+})();
